Test for get tracer function with addition functions and fixture
